+++ NEWS.md
@@ -7,12 +7,14 @@
 * Add a `characters` argument of `tar_traceback()` to cap the traceback line lengths (#383).
 * Add new "summary" and "about" views to `tar_watch()` (#382).
 * Implement `tar_poll()` to repeatedly poll runtime progress in the R console (#381). `tar_poll()` is a lightweight alternative to `tar_watch()`.
+* Change the color of the "dormant" status in the graph.
 
 ## Enhancements
 
 * Attempt to print only the useful part of the traceback in `tar_traceback()` (#383).
 * Right-align the headers and counts in the "summary" and "forecast" reporters.
 * Add a line break at the end of the "summary" reporter so warnings do not mangle the output.
+* In `tar_watch()`, use `shinybusy` instead of `shinycssloaders` and keep current output on display while new output is rendering (#386, @rcorty).
 
 # targets 0.3.1